# docs/swagger.yml
...
openapi: 3.0.0
info:
  title: EasyTex API
  description: Документация по API EasyTex
  version: 0.1.9
servers:
  - url: http://localhost:8080/api/v1
    description: Localhost server
  - url: https://smartlectures.ru/api/v1
    description: Main server
paths:
...
  /auth/login:
    post:
      summary: Авторизация пользователя
      responses:
        '204':
          description: Ответ при успешном поиске
          headers: 
            Set-Cookie:
              schema: 
                type: string
                example: session=abcde12345; Path=/; HttpOnly
        '400':
          description: Ответ, при неверной почте или пароле
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ответ при ошибке на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      tags: [auth]
  /auth/register:
    put:
      summary: Регистрация пользователя
      responses:
        '200':
          description: Ответ при успешном поиске
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Snippet'
        '409':
          description: Ответ, если такой пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ответ при ошибке на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      tags: [auth]
  /auth/logout:
    delete:
      summary: Удаление текущей авторизационной сессии
      responses:
        '204':
          description: Ответ при успешном удалении сессии
        '500':
          description: Ответ при ошибке на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags: [auth]
  /auth/me:
    get:
      summary: Получение информации о текущем пользователе
      responses:
        '200':
          description: Ответ при успешном поиске
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Ответ для не авторизованного пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ответ при ошибке на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags: [auth]
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Snippet:
      type: object
      properties:
        snippetId:
          type: number
          format: int32
          example: 1
          x-isnullable: false
        name:
          type: string
          description: Название сниппета
          example: integral
          x-isnullable: false
        description:
          type: string
          description: Описание сниппета
          example: Определенный интеграл
          x-isnullable: false
        body:
          type: string
          description: Тело сниппета
          example: '```$$\alpha$$```'
          x-isnullable: false
        userId:
          type: number
          description: ID владельца сниппета
          format: int32
          example: 1
          x-isnullable: false
      required: [snippetId, name, description, body, userId]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          description: Электронная почта
          example: default@default.ru
          x-isnullable: false
        password:
          type: string
          description: Пароль
          example: secret
          x-isnullable: false
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          description: Электронная почта
          example: default@default.ru
          x-isnullable: false
        password:
          type: string
          description: Пароль
          example: secret
          x-isnullable: false
        name:
          type: string
          description: имя пользователя
          example: name
          x-isnullable: false
        surname:
          type: string
          description: фамилия пользователя
          example: surname
          x-isnullable: false
        username:
          type: string
          description: ник пользователя
          example: username
          x-isnullable: false
      required: [email, password, username]
...
    UserInfo:
      type: object
      properties:
        email:
          type: string
          description: Электронная почта
          example: default@default.ru
          x-isnullable: false
        password:
          type: string
          description: Пароль
          example: secret
          x-isnullable: false
        name:
          type: string
          description: имя пользователя
          example: name
          x-isnullable: false
        surname:
          type: string
          description: фамилия пользователя
          example: surname
          x-isnullable: false
        username:
          type: string
          description: ник пользователя
          example: username
          x-isnullable: false
      required: [email, password, username]
...
    Error:
      type: object
      properties:
        msg:
          type: string
          description: Текстовое описание ошибки
          example: note with this id not found
      required: [msg]
...